on the way to depreceating model/nb and to improving the hierarchy accessability of models
