+++ batchglm/api/models/nb_glm.py
@@ -1,12 +1,4 @@
-from batchglm.models.nb_glm.base import INPUT_DATA_PARAMS
-from batchglm.models.nb_glm.base import MODEL_PARAMS
-from batchglm.models.nb_glm.base import ESTIMATOR_PARAMS
-from batchglm.models.nb_glm.base import InputData
-from batchglm.models.nb_glm.base import Model
-# from batchglm.models.nb_glm.base import AbstractEstimator
-# from batchglm.models.nb_glm.base import model_from_params
-
+from batchglm.models.nb_glm.input import InputData
+from batchglm.models.nb_glm.model import Model
 from batchglm.models.nb_glm.simulator import Simulator
-
-# use TF as default estimator implementation
 from batchglm.train.tf.nb_glm.estimator import Estimator
