+++ batchglm/models/external.py
@@ -1,3 +1,1 @@
-import batchglm.utils.random as rand_utils
-import batchglm.data as data_utils
 import batchglm.pkg_constants as pkg_constants
