+++ batchglm/models/glm/__init__.py
@@ -1,3 +1,3 @@
-from .base import Model
+from .base import BasicGLM
 from .utils import parse_design
 from .utils import closedform_glm_mean, closedform_glm_var
