+++ batchglm/models/glm/external.py
@@ -1,0 +1,2 @@
+from batchglm.utils.linalg import groupwise_solve_lm
+from batchglm.utils.numeric import weighted_mean, weighted_variance
