+++ batchglm/models/nb_glm/__init__.py
@@ -1,0 +1,2 @@
+from .model import Model, XArrayModel
+from .input import InputData
