+++ batchglm/models/nb_glm/input.py
@@ -1,4 +1,3 @@
-import abc
 from typing import Union
 
 try:
@@ -9,44 +8,20 @@ import xarray as xr
 import numpy as np
 import pandas as pd
 
-from ..nb.base import InputData as NegativeBinomialInputData
-from ..nb.base import Model as NegativeBinomialModel
-from ..nb.base import MODEL_PARAMS as NB_MODEL_PARAMS
-from ..nb.base import INPUT_DATA_PARAMS as NB_INPUT_DATA_PARAMS
-from ..base import BasicEstimator
-from ..glm import Model as GeneralizedLinearModel
-from ..glm import parse_design
+from .external import BasicInputData, parse_design
 
 import patsy
 
-INPUT_DATA_PARAMS = NB_INPUT_DATA_PARAMS.copy()
-INPUT_DATA_PARAMS.update({
+INPUT_DATA_PARAMS = {
+    "X": ("observations", "features"),
     "design_loc": ("observations", "design_loc_params"),
     "design_scale": ("observations", "design_scale_params"),
+    "constraints_loc": ("constraints_loc_model", "design_loc_params"),
+    "constraints_scale": ("constraints_scale_model", "design_scale_params"),
     "size_factors": ("observations",),
-})
+}
 
-MODEL_PARAMS = NB_MODEL_PARAMS.copy()
-MODEL_PARAMS.update(INPUT_DATA_PARAMS)
-MODEL_PARAMS.update({
-    "a": ("design_loc_params", "features"),
-    "b": ("design_scale_params", "features"),
-    "par_link_loc": ("design_loc_params", "features"),
-    "par_link_scale": ("design_scale_params", "features"),
-    # "par_loc": ("design_loc_params", "features"),
-    # "par_scale": ("design_scale_params", "features"),
-})
-
-ESTIMATOR_PARAMS = MODEL_PARAMS.copy()
-ESTIMATOR_PARAMS.update({
-    "loss": (),
-    "gradient": ("features",),
-    "hessians": ("features", "delta_var0", "delta_var1"),
-    "fisher_inv": ("features", "delta_var0", "delta_var1"),
-})
-
-
-class InputData(NegativeBinomialInputData):
+class InputData(BasicInputData):
     """
     Input data for Generalized Linear Models (GLMs) with negative binomial noise.
     """
@@ -154,7 +129,6 @@ class InputData(NegativeBinomialInputData):
         if cast_dtype is not None:
             design_loc = design_loc.astype(cast_dtype)
             design_scale = design_scale.astype(cast_dtype)
-            # design = design.chunk({"observations": 1})
 
         retval.design_loc = design_loc
         retval.design_scale = design_scale
@@ -177,11 +151,11 @@ class InputData(NegativeBinomialInputData):
 
     @property
     def constraints_loc(self) -> np.ndarray:
-        return self._constraints_loc
+        return self.data["constraints_loc"]
 
     @constraints_loc.setter
     def constraints_loc(self, data: np.ndarray):
-        self._constraints_loc = data
+        self.data["constraints_loc"] = data
 
     @property
     def design_loc_names(self) -> xr.DataArray:
@@ -201,11 +175,11 @@ class InputData(NegativeBinomialInputData):
 
     @property
     def constraints_scale(self) -> np.ndarray:
-        return self._constraints_scale
+        return self.data["constraints_scale"]
 
     @constraints_scale.setter
     def constraints_scale(self, data: np.ndarray):
-        self._constraints_scale = data
+        self.data["constraints_scale"] = data
 
     @property
     def design_scale_names(self) -> xr.DataArray:
@@ -256,251 +230,3 @@ class InputData(NegativeBinomialInputData):
         super().set_chunk_size(cs)
         self.design_loc = self.design_loc.chunk({"observations": cs})
         self.design_scale = self.design_scale.chunk({"observations": cs})
-
-
-class Model(GeneralizedLinearModel, NegativeBinomialModel, metaclass=abc.ABCMeta):
-    """
-    Generalized Linear Model (GLM) with negative binomial noise.
-    Uses the natural logarithm as linker function.
-    """
-
-    @classmethod
-    def param_shapes(cls) -> dict:
-        return MODEL_PARAMS
-
-    @property
-    @abc.abstractmethod
-    def input_data(self) -> InputData:
-        pass
-
-    @property
-    def design_loc(self) -> xr.DataArray:
-        return self.input_data.design_loc
-
-    @property
-    def design_scale(self) -> xr.DataArray:
-        return self.input_data.design_scale
-
-    @property
-    def size_factors(self) -> Union[xr.DataArray, None]:
-        return self.input_data.size_factors
-
-    @property
-    @abc.abstractmethod
-    def a(self) -> xr.DataArray:
-        pass
-
-    @property
-    @abc.abstractmethod
-    def b(self) -> xr.DataArray:
-        pass
-
-    @property
-    def mu(self) -> xr.DataArray:
-        # exp(design * a + sf)
-        # Depend on xarray and use xr.DataArray.dot() for matrix multiplication
-        # Also, make sure that the right order of dimensions get returned => use xr.DataArray.transpose()
-        log_retval = self.design_loc.dot(self.par_link_loc, dims="design_loc_params")
-        log_retval = log_retval.transpose(*self.param_shapes()["mu"])
-
-        if self.size_factors is not None:
-            log_retval += self.link_loc(self.size_factors)
-
-        return np.exp(log_retval)
-
-    @property
-    def r(self) -> xr.DataArray:
-        # exp(design * b + sf)
-        # Depend on xarray and use xr.DataArray.dot() for matrix multiplication
-        # Also, make sure that the right order of dimensions get returned => use xr.DataArray.transpose()
-        log_retval = self.design_scale.dot(self.par_link_scale, dims="design_scale_params")
-        log_retval = log_retval.transpose(*self.param_shapes()["r"])
-
-        if self.size_factors is not None:
-            log_retval += self.link_loc(self.size_factors)
-
-        return np.exp(log_retval)
-
-    @property
-    def par_link_loc(self):
-        return self.a
-
-    @property
-    def par_link_scale(self):
-        return self.b
-
-    def link_loc(self, data):
-        return np.log(data)
-
-    def inverse_link_loc(self, data):
-        return np.exp(data)
-
-    def link_scale(self, data):
-        return np.log(data)
-
-    def inverse_link_scale(self, data):
-        return np.exp(data)
-
-    @property
-    def location(self):
-        return self.mu
-
-    @property
-    def scale(self):
-        return self.r
-
-    def export_params(self, append_to=None, **kwargs):
-        if append_to is not None:
-            if isinstance(append_to, anndata.AnnData):
-                # append_to.obsm["design"] = self.design
-                append_to.varm["a"] = np.transpose(self.a)
-                append_to.varm["b"] = np.transpose(self.b)
-            elif isinstance(append_to, xr.Dataset):
-                # append_to["design"] = (self.param_shapes()["design"], self.design)
-                append_to["a"] = (self.param_shapes()["a"], self.a)
-                append_to["b"] = (self.param_shapes()["b"], self.b)
-            else:
-                raise ValueError("Unsupported data type: %s" % str(type(append_to)))
-        else:
-            ds = xr.Dataset({
-                # "design": (self.param_shapes()["design"], self.design),
-                "a": (self.param_shapes()["a"], self.a),
-                "b": (self.param_shapes()["b"], self.b),
-            })
-            return ds
-
-
-def _model_from_params(data: Union[xr.Dataset, anndata.AnnData, xr.DataArray], params=None, a=None, b=None):
-    input_data = InputData.new(data)
-
-    if params is None:
-        if isinstance(data, Model):
-            params = xr.Dataset({
-                "a": data.a,
-                "b": data.b,
-            })
-        elif anndata is not None and isinstance(data, anndata.AnnData):
-            params = xr.Dataset({
-                "a": (MODEL_PARAMS["a"], np.transpose(data.varm["a"])),
-                "b": (MODEL_PARAMS["b"], np.transpose(data.varm["b"])),
-            })
-        elif isinstance(data, xr.Dataset):
-            params = data
-        else:
-            params = xr.Dataset({
-                "a": (MODEL_PARAMS["a"], a) if not isinstance(a, xr.DataArray) else a,
-                "b": (MODEL_PARAMS["b"], b) if not isinstance(b, xr.DataArray) else b,
-            })
-
-    return input_data, params
-
-
-def model_from_params(*args, **kwargs) -> Model:
-    (input_data, params) = _model_from_params(*args, **kwargs)
-    return XArrayModel(input_data, params)
-
-
-class XArrayModel(Model):
-    _input_data: InputData
-    params: xr.Dataset
-
-    def __init__(self, input_data: InputData, params: xr.Dataset):
-        self._input_data = input_data
-        self.params = params
-
-    @property
-    def input_data(self) -> InputData:
-        return self._input_data
-
-    @property
-    def a(self):
-        return self.params['a']
-
-    @property
-    def b(self):
-        return self.params['b']
-
-    def __str__(self):
-        return "[%s.%s object at %s]: data=%s" % (
-            type(self).__module__,
-            type(self).__name__,
-            hex(id(self)),
-            self.params
-        )
-
-    def __repr__(self):
-        return self.__str__()
-
-
-# class AnndataModel(Model):
-#     data: anndata.AnnData
-#
-#     def __init__(self, data: anndata.AnnData):
-#         self.data = data
-#
-#     @property
-#     def X(self):
-#         return self.data.X
-#
-#     @property
-#     def design(self):
-#         return self.data.obsm["design"]
-#
-#     @property
-#     def a(self):
-#         return np.transpose(self.data.varm["a"])
-#
-#     @property
-#     def b(self):
-#         return np.transpose(self.data.varm["b"])
-
-
-class AbstractEstimator(Model, BasicEstimator, metaclass=abc.ABCMeta):
-
-    @classmethod
-    def param_shapes(cls) -> dict:
-        return ESTIMATOR_PARAMS
-
-
-class XArrayEstimatorStore(AbstractEstimator, XArrayModel):
-
-    def initialize(self, **kwargs):
-        raise NotImplementedError("This object only stores estimated values")
-
-    def train(self, **kwargs):
-        raise NotImplementedError("This object only stores estimated values")
-
-    def finalize(self, **kwargs) -> AbstractEstimator:
-        return self
-
-    def validate_data(self, **kwargs):
-        raise NotImplementedError("This object only stores estimated values")
-
-    def __init__(self, estim: AbstractEstimator):
-        input_data = estim.input_data
-        # to_xarray triggers the get function of these properties and thereby
-        # causes evaluation of the properties that have not been computed during
-        # training, such as the hessian.
-        params = estim.to_xarray(["a", "b", "loss", "gradient", "hessians", "fisher_inv"], coords=input_data.data)
-
-        XArrayModel.__init__(self, input_data, params)
-
-    @property
-    def input_data(self):
-        return self._input_data
-
-    @property
-    def loss(self):
-        return self.params["loss"]
-
-    @property
-    def gradient(self):
-        return self.params["gradient"]
-
-    @property
-    def hessians(self):
-        return self.params["hessians"]
-
-    @property
-    def fisher_inv(self):
-        return self.params["fisher_inv"]
